图片资源映射bug修复
