+++ src/main/java/com/example/gamevibe/config/WebMvcConfig.java
@@ -2,12 +2,15 @@ package com.example.gamevibe.config;
 
 import com.example.gamevibe.interceptor.IdempotentInterceptor;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import javax.annotation.Resource;
 
 @Configuration
+@EnableWebMvc
 public class WebMvcConfig implements WebMvcConfigurer {
 
     @Resource
@@ -19,5 +22,15 @@ public class WebMvcConfig implements WebMvcConfigurer {
                 .addPathPatterns("/api/post_like/like")
                 .addPathPatterns("/api/post_collect/collect")
                 .addPathPatterns("/api/focus_user/focus");
+    }
+
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("/imgs/**")
+                .addResourceLocations("file:/www/wwwroot/images/");
+        registry.addResourceHandler("doc.html")
+                .addResourceLocations("classpath:/META-INF/resources/");
+        registry.addResourceHandler("/webjars/**")
+                .addResourceLocations("classpath:/META-INF/resources/webjars/");
     }
 }
