+++ src/main/resources/application.yml
@@ -79,4 +79,5 @@ casdoor:
   application-name: app-built-in
   redirect-url: http://localhost:8080/api/signin
 gamevibe:
-  databasePath: http://localhost:8080/imgs/
+  databasePath: http://8.138.154.250:8080/imgs/
+  serverPath: /www/wwwroot/images/
